Finished frame extractor, added MSE,  added two picking frames functions, trying to deal with noise
Picking functions pick frames without a cat too. I thought it was because of time, but turns out there is a lot of noise that builds up as video goes. The cat also affects it in some way (changes lightning/focus/compression algo in the camera). Tried multiple denoising methods with partial success